Change of plan: ares should be folded into higan, not the other way around.
Originally I was planning to archive higan and try to move everything over
to ares, since it was newer. However, higan has been around for years and
ares only a few months, and ares only ever saw one publicised release, so as
a matter of practicality we should probably fold it back into higan.

As a result, I've removed or modified wording that suggested ares was the
new hotness, and that higan was old and busted.